+++ access-control-mongoose/lib/decorators/index.ts
@@ -2,3 +2,4 @@ export * from "./created-policy.decorator";
 export * from "./deleted-policy.decorator";
 export * from "./policy.decorator";
 export * from "./updated-policy.decorator";
+export * from "./access-control-resource.decorator";