+++ access-control/lib/deserializers/user.deserializer.ts
@@ -1,5 +1,5 @@
 import { Users } from "../models";
 
 export abstract class UserDeserializer {
-    public abstract deserializeUser(req: unknown): Promise<Users>;
+    public abstract deserializeUser(req: unknown): Promise<Users | null>;
 }