+++ data-filter/lib/filter/controllers/filter.controller.ts
@@ -70,7 +70,7 @@ export class FilterController<Data> {
     public async searchFilterResourceValue(
         @Query() search: FilterResourceValueModel,
         @Req() req: any
-    ): Promise<SelectFilterValue> {
+    ): Promise<SelectFilterValue | null> {
         return await this.filterService.findResourceValueById(req, search, await this.getUser(req) ?? {} as DataFilterUserModel);
     }