+++ data-filter/lib/sequelize.utils.ts
@@ -21,7 +21,7 @@ export interface GeoPoint {
 export class M extends Model {}
 
 export class SequelizeUtils {
-    public static reduceIncludes(includes: IncludeOptions[][], ignoreAttributes = false): Includeable[] {
+    public static reduceIncludes(includes: Array<IncludeOptions | IncludeOptions[]>, ignoreAttributes = false): Includeable[] {
         let include: IncludeOptions[] = [];
         for (const i of includes) {
             include = this.mergeIncludes(include, i, ignoreAttributes);
@@ -30,7 +30,15 @@ export class SequelizeUtils {
         return include;
     }
 
-    public static mergeIncludes(a: IncludeOptions[] = [], b: IncludeOptions[] = [], ignoreAttributes = false) {
+    public static mergeIncludes(a: IncludeOptions | IncludeOptions[] = [], b: IncludeOptions | IncludeOptions[] = [], ignoreAttributes = false) {
+        if (!Array.isArray(a)) {
+            a = [a];
+        }
+
+        if (!Array.isArray(b)) {
+            b = [b];
+        }
+
         for (const bChild of b) {
             const aChild = a.find(value => value.model === bChild.model && value.as === bChild.as);
             if (aChild) {
@@ -55,6 +63,9 @@ export class SequelizeUtils {
                 }
                 if (aChild.separate || bChild.separate) {
                     aChild.separate = true;
+                }
+                if (bChild.limit) {
+                    aChild.limit ??= bChild.limit;
                 }
                 aChild.paranoid = !(aChild.paranoid === false || bChild.paranoid === false);
                 if (TypeUtils.isNotNullOrUndefined(aChild.subQuery) || TypeUtils.isNotNullOrUndefined(bChild.subQuery)) {
@@ -148,7 +159,7 @@ export class SequelizeUtils {
         if (attribute.exclude) {
             return {
                 exclude: attribute.exclude.filter((x) => x !== "id")
-            }
+            };
         }
 
         return attribute;
@@ -201,16 +212,16 @@ export class SequelizeUtils {
                 return { [Op.notIn]: rule.value as any[] };
 
             case "less":
-                return { [Op.lt]: rule.value as any};
+                return { [Op.lt]: rule.value as any };
 
             case "less_or_equal":
-                return { [Op.lte]: rule.value as any};
+                return { [Op.lte]: rule.value as any };
 
             case "greater":
-                return { [Op.gt]: rule.value as any};
+                return { [Op.gt]: rule.value as any };
 
             case "greater_or_equal":
-                return { [Op.gte]: rule.value as any};
+                return { [Op.gte]: rule.value as any };
 
             case "between":
                 return { [Op.between]: rule.value as any };
@@ -287,15 +298,14 @@ export class SequelizeUtils {
                 }
 
                 return !["DATE", "DATEONLY", "VIRTUAL"].some(t => t === (attr.type as AbstractDataTypeConstructor).key);
-            }).
-            map(x => {
+            }).map(x => {
                 const attr = attributes[x];
                 if (!attr) {
                     return x;
                 }
 
                 return attr.field ? attr.field : x;
-            })
+            });
     }
 
     public static reduceModelFromPath(model: M, path: string) {